refactor(name) Name을 global 레벨로 분리
