+++ src/main/java/kitchenpos/product/application/port/in/ProductRegistrationUseCase.java
@@ -2,8 +2,8 @@ package kitchenpos.product.application.port.in;
 
 import kitchenpos.product.application.ProductDTO;
 import kitchenpos.product.application.exception.ContainsProfanityException;
-import kitchenpos.product.domain.Name;
 import kitchenpos.product.domain.ProductPrice;
+import kitchenpos.support.vo.Name;
 
 public interface ProductRegistrationUseCase {
