+++ src/main/java/kitchenpos/product/domain/ProductNameFactory.java
@@ -1,7 +1,9 @@
 package kitchenpos.product.domain;
 
+import kitchenpos.support.vo.Name;
+
 public interface ProductNameFactory {
-	
+
     /**
      * @throws IllegalArgumentException nameCandidate에 비속어가 포함되어 있을 때
      */
