+++ src/test/java/kitchenpos/product/application/DefaultProductRegistrationUseCaseTest.java
@@ -9,11 +9,11 @@ import kitchenpos.product.Fixtures;
 import kitchenpos.product.application.exception.ContainsProfanityException;
 import kitchenpos.product.application.port.in.ProductRegistrationUseCase;
 import kitchenpos.product.application.port.out.ProductNewRepository;
-import kitchenpos.product.domain.Name;
 import kitchenpos.product.domain.ProductNameAccessor;
 import kitchenpos.product.domain.ProductNameFactory;
 import kitchenpos.product.domain.ProductPrice;
 import kitchenpos.product.fakerepository.ProductNewFakeRepository;
+import kitchenpos.support.vo.Name;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator.ReplaceUnderscores;
