+++ src/test/java/kitchenpos/product/domain/ProductNameAccessor.java
@@ -1,5 +1,7 @@
 package kitchenpos.product.domain;
 
+import kitchenpos.support.vo.Name;
+
 public final class ProductNameAccessor {
 
     public static ProductName create(final Name name) {
